style: :art: imporve code lisibility for manager

+++ naas/manager.py
@@ -194,27 +194,38 @@ class Manager:
         else:
             raise FileNotFoundError(f"File {path} not Found")
 
-    def get_prod(self, path):
-        self.__copy_file_in_dev(path)
-
     def get_out_path(self, path):
         filename = os.path.basename(path)
         dirname = os.path.dirname(path)
         out_path = os.path.join(dirname, f"out_{filename}")
         return out_path
 
+    def get_prod(self, path):
+        current_file = self.get_path(path)
+        self.__copy_file_in_dev(current_file)
+        print(
+            "🕣 Your Notebook from production folder has been copied into your dev folder.\n"
+        )
+
     def get_output(self, path):
         out_path = self.get_out_path(path)
         self.__copy_file_in_dev(out_path)
+        print(
+            "🕣 Your Notebook OUTPUT from production folder has been copied into your dev folder\n"
+        )
 
     def clear_output(self, path):
         out_path = self.get_out_path(path)
-        print(f"Delete {out_path}")
-        os.remove(out_path)
+        if os.path.exists(out_path):
+            os.remove(out_path)
+            print("🕣 Your Notebook output has been remove from production folder.\n")
+        else:
+            raise FileNotFoundError(f"File {out_path} not Found")
 
     def list_history(self, path):
-        prod_path = self.get_prod_path(path)
-        filename = os.path.basename(path)
+        current_file = self.get_path(path)
+        prod_path = self.get_prod_path(current_file)
+        filename = os.path.basename(current_file)
         dirname = os.path.dirname(prod_path)
         print("Avaliable :\n")
         for ffile in os.listdir(dirname):
@@ -228,15 +239,19 @@ class Manager:
                 print(histo + "\n")
 
     def get_history(self, path, histo):
-        filename = os.path.basename(path)
-        dirname = os.path.dirname(path)
+        current_file = self.get_path(path)
+        filename = os.path.basename(current_file)
+        dirname = os.path.dirname(current_file)
         path_histo = os.path.join(dirname, f"{histo}_{filename}")
         self.__copy_file_in_dev(path_histo)
 
     def clear_history(self, path, histo=None):
-        prod_path = self.get_prod_path(path)
+        current_file = self.get_path(path)
+        prod_path = self.get_prod_path(current_file)
         filename = (
-            os.path.basename(path) if not histo else f"{histo}_{os.path.basename(path)}"
+            os.path.basename(current_file)
+            if not histo
+            else f"{histo}_{os.path.basename(current_file)}"
         )
         dirname = os.path.dirname(prod_path)
         for ffile in os.listdir(dirname):
